requirements.txt
datasets==3.6.0
numpy==1.26.4
pandas==2.0.3
peft==0.16.0
PyYAML==6.0.2
scikit-learn==1.7.0
transformers==4.53.1
trl==0.19.0
tensorboardx==2.6.4